requirements.txt
django
south
psycopg2
djangorestframework
